+++ src/middlewares/middleware-activation-condition/filter-realities.ts
@@ -1,9 +1,13 @@
 import { PolarisRequestHeaders } from '@enigmatis/utills';
+import { PolarisContext } from '../../server/polaris-context';
 import { ResponseMiddlewareParams } from '../middleware';
 import { MiddlewareCondition } from './filter-condition';
 
 class FilterRealities implements MiddlewareCondition {
-    shouldBeReturned({ context, result }: ResponseMiddlewareParams, isSubEntity: boolean): boolean {
+    shouldBeReturned(
+        { context, result }: { context: PolarisContext; result: any },
+        isSubEntity: boolean,
+    ): boolean {
         const operationalRealityId: number = 0;
         const headers: PolarisRequestHeaders = context.headers;
         const realityIdHeaderExists: any = headers.realityId === 0 ? true : headers.realityId;